Restructure package.yaml
