+++ tests/PrimeFieldTests.hs
@@ -3,6 +3,7 @@ module PrimeFieldTests where
 import Protolude
 
 import PrimeField
+import Test.Tasty
 
 import GaloisFieldTests
 
@@ -21,26 +22,41 @@ type FVL = PrimeField 20988936657440586486151264256610222593863921
 type FXL = PrimeField 5210644015679228794060694325390955853335898483908056458352183851018372555735221
 type FZL = PrimeField 741640062627530801524787141901937474059940781097519023905821316144415759504705008092818711693940737
 
+type Fq = PrimeField 21888242871839275222246405745257275088696311157297823662689037894645226208583
+
+
+test_S2 :: TestTree
 test_S2 = fieldAxioms (Proxy :: Proxy FS2) "FS2"
 
+test_S3 :: TestTree
 test_S3 = fieldAxioms (Proxy :: Proxy FS3) "FS3"
 
+test_S5 :: TestTree
 test_S5 = fieldAxioms (Proxy :: Proxy FS5) "FS5"
 
+test_S7 :: TestTree
 test_S7 = fieldAxioms (Proxy :: Proxy FS7) "FS7"
 
+test_M0 :: TestTree
 test_M0 = fieldAxioms (Proxy :: Proxy FM0) "FM0"
 
+test_M1 :: TestTree
 test_M1 = fieldAxioms (Proxy :: Proxy FM1) "FM1"
 
+test_M2 :: TestTree
 test_M2 = fieldAxioms (Proxy :: Proxy FM2) "FM2"
 
+test_M3 :: TestTree
 test_M3 = fieldAxioms (Proxy :: Proxy FM3) "FM3"
 
+test_M4 :: TestTree
 test_M4 = fieldAxioms (Proxy :: Proxy FM4) "FM4"
 
+test_VL :: TestTree
 test_VL = fieldAxioms (Proxy :: Proxy FVL) "FVL"
 
+test_XL :: TestTree
 test_XL = fieldAxioms (Proxy :: Proxy FXL) "FXL"
 
+test_ZL :: TestTree
 test_ZL = fieldAxioms (Proxy :: Proxy FZL) "FZL"
